+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `--quiet` flag to suppress any non-error output to stderr
 - Fixed a bug for the `--restrict-collaboration` flag for `box folders:update` where previously the flag would not restrict the collaborations when passed as true and would restrict collaborations when passed as false
 - Added `box trash:restore` to restore a trashed item and `box trash:get` to get information on a trashed item
+- Fixed a bug where flags that can be specified multiple times in a single command could not be passed through the command line for bulk commands
 
 ## 2.4.0 [2019-08-29]